+++ Source/Core/BaseRequest.swift
@@ -71,24 +71,13 @@ public class BaseRequest<Model: ResponseParseable, ErrorModel: ResponseParseable
     public let path: String
     
     /// HTTP method
-    public var method: Alamofire.Method = .GET {
-        didSet {
-            encoding = encodingStrategy(method)
-        }
-    }
+    public var method: Alamofire.Method = .GET
     
     /// Parameters of current request.
     public var parameters: [String: AnyObject] = [:]
     
     /// Selection of encoding based on HTTP method.
     public var encodingStrategy : (Alamofire.Method) -> Alamofire.ParameterEncoding
-    
-    /// Parameter encoding option.
-    /// This property is deprecated and may be removed in following releases.
-    /// Please use TRON.encodingStrategy property to set strategy globally, or 
-    /// APIRequest encodingStrategy property to set it locally for single request.
-    @available(*, deprecated, message: "Use encodingStrategy property on TRON or APIRequest.")
-    public var encoding: Alamofire.ParameterEncoding = .url
     
     /// Headers, that should be used for current request.
     /// - Note: Resulting headers may include global headers from `TRON` instance and `Alamofire.Manager` defaultHTTPHeaders.